+++ src/com/kaylerrenslow/armaDialogCreator/control/sv/SVNumber.java
@@ -5,7 +5,7 @@ package com.kaylerrenslow.armaDialogCreator.control.sv;
  @since Created on 07/15/2016.
  @see SVInteger
  @see SVDouble */
-public abstract class SVNumber extends SerializableValue {
+public abstract class SVNumber extends SerializableValue implements SVNumericValue {
 	public SVNumber(Number value) {
 		this(value.toString() + "");
 	}