+++ src/com/kaylerrenslow/armaDialogCreator/control/sv/SVNumericValue.java
@@ -1,0 +1,10 @@
+package com.kaylerrenslow.armaDialogCreator.control.sv;
+
+/**
+ @author Kayler
+ @since 07/04/2017 */
+public interface SVNumericValue {
+	int toInt();
+
+	double toDouble();
+}